reboot-website/requirements.txt
alembic==0.8.6
Flask==0.10.1
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
python-editor==1.0
SQLAlchemy==1.0.12
Werkzeug==0.11.9
request==0.0.5
